Update CI setup in general and add more detailed explanations into docs. (#49)

+++ scripts/setup-repository-ci.bash
@@ -15,13 +15,11 @@
 # limitations under the License.
 
 
-usage='setup-ci-config.bash "repo_name" "repo_namespace"'
-#
+usage='setup-repository-ci.bash "repo_name" "repo_namespace" ["first_package second_package ..."]'
 
 # Load Framework defines
 script_own_dir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" > /dev/null && pwd )"
 source $script_own_dir/../setup.bash
-check_and_set_ros_distro_and_version ${ROS_DISTRO}
 
 repo_name=$1
 if [ -z "$1" ]; then
@@ -43,7 +41,7 @@ read -p "${RAW_TERMINAL_COLOR_BROWN}Are you setting CI for private repository?${
 private=${private:="no"}
 
 CI_FORMAT="ci-format"
-if  [[ "$choice" == "yes" ]]; then
+if  [[ "$private" == "yes" ]]; then
   CI_FORMAT="ci-format-private"
 fi
 
@@ -55,8 +53,10 @@ CI_REUSABLE_ROS_TOOLING="reusable-ros-tooling-source-build"
 PRE_RELEASE="prerelease-check"
 
 CI_ABI_COMPATIBILITY="abi-compatibility"
-CI_BINARY_BUILD="binary-build"
-CI_SEMI_BINARY_BUILD="semi-binary-build"
+CI_BINARY_BUILD_MAIN="binary-build-main"
+CI_BINARY_BUILD_TESTING="binary-build-testing"
+CI_SEMI_BINARY_BUILD_MAIN="semi-binary-build-main"
+CI_SEMI_BINARY_BUILD_TESTING="semi-binary-build-testing"
 CI_SOURCE_BUILD="source-build"
 CI_RHEL_BINARY_BUILD="rhel-binary-build"
 
@@ -73,14 +73,18 @@ CI_GENERAL_FILES=(
 
 CI_DISTRIBUTION_FILES=(
   $CI_ABI_COMPATIBILITY
-  $CI_BINARY_BUILD
-  $CI_SEMI_BINARY_BUILD
+  $CI_BINARY_BUILD_MAIN
+  $CI_BINARY_BUILD_TESTING
+  $CI_SEMI_BINARY_BUILD_MAIN
+  $CI_SEMI_BINARY_BUILD_TESTING
   $CI_SOURCE_BUILD
   $CI_RHEL_BINARY_BUILD
 )
 
-read -p "${RAW_TERMINAL_COLOR_BROWN}Name of the default ROS distro?${RAW_TERMINAL_COLOR_NC} [default: rolling] " default_ros_distro
+read -p "${RAW_TERMINAL_COLOR_BROWN}Name of the default ROS distro? Supported distributions(${rtw_supported_ros_distributions[*]})${RAW_TERMINAL_COLOR_NC} [default: rolling] " default_ros_distro
 default_ros_distro=${default_ros_distro:="rolling"}
+check_and_set_ros_distro_and_version "$default_ros_distro"
+default_ros_distro=${ros_distro}
 read -p "${RAW_TERMINAL_COLOR_BROWN}Name of the repository default branch?${RAW_TERMINAL_COLOR_NC} [default: master]: " default_branch
 default_branch=${default_branch:="master"}
 
@@ -90,7 +94,7 @@ mkdir -p .github/workflows
 # Coping and SEDing general CI files
 for CI_FILE in "${CI_GENERAL_FILES[@]}"; do
   cp -n ${PACKAGE_TEMPLATES}/CI-github_${CI_FILE}.yml .github/workflows/${CI_FILE}.yml
-  sed -i 's/\$ROS_DISTRO\$/'${default_ros_distro}'/g' .github/workflows/${CI_FILE}.yml
+  sed -i 's/\$ros_distro\$/'${default_ros_distro}'/g' .github/workflows/${CI_FILE}.yml
   sed -i 's/\$branch\$/'${default_branch}'/g' .github/workflows/${CI_FILE}.yml
   sed -i 's/\$NAME\$/'${repo_name}'/g' .github/workflows/${CI_FILE}.yml
   if [[ "$package_names" != "<none>" ]]; then
@@ -112,13 +116,15 @@ some_ros_distro=${default_ros_distro}
 some_branch=${default_branch}
 
 TMP_FILE=".f_tmp"
+TMP_FILE_2=".f_tmp2"
 touch $TMP_FILE
+touch $TMP_FILE_2
 
 while true; do
   for CI_FILE in "${CI_DISTRIBUTION_FILES[@]}"; do
     output_file_name="${some_ros_distro}-${CI_FILE}"
     cp -n ${PACKAGE_TEMPLATES}/CI-github_${CI_FILE}.yml .github/workflows/${output_file_name}.yml
-    sed -i 's/\$ROS_DISTRO\$/'${some_ros_distro}'/g' .github/workflows/${output_file_name}.yml
+    sed -i 's/\$ros_distro\$/'${some_ros_distro}'/g' .github/workflows/${output_file_name}.yml
     sed -i 's/\$Ros_distro\$/'${some_ros_distro^}'/g' .github/workflows/${output_file_name}.yml
     sed -i 's/\$branch\$/'${some_branch}'/g' .github/workflows/${output_file_name}.yml
     sed -i 's/\$NAME\$/'${repo_name}'/g' .github/workflows/${output_file_name}.yml
@@ -149,17 +155,30 @@ while true; do
   head -$CUT_LINE README.md >> $TMP_FILE
   cat ${PACKAGE_TEMPLATES}/_append_to_README_build_status_table.md >> $TMP_FILE
 
-  sed -i 's/\$branch\$/'${some_branch}'/g' $TMP_FILE
-  sed -i 's/\$ros_distro\$/'${some_ros_distro}'/g' $TMP_FILE
-  sed -i 's/\$Ros_distro\$/'${some_ros_distro^}'/g' $TMP_FILE
-  sed -i 's/\$NAME\$/'${repo_name}'/g' $TMP_FILE
-  sed -i 's/\$NAMESPACE\$/'${repo_namespace}'/g' $TMP_FILE
+  # Update .github/workflows/README.md file
+  cat ${PACKAGE_TEMPLATES}/_append_to_workflows_README_build_status_table.md > $TMP_FILE_2
 
+  # SED temp files
+  FILES_TO_UPDATE=("$TMP_FILE" "$TMP_FILE_2")
+  for SED_FILE in "${FILES_TO_UPDATE[@]}"; do
+    sed -i 's/\$branch\$/'${some_branch}'/g' $SED_FILE
+    sed -i 's/\$ros_distro\$/'${some_ros_distro}'/g' $SED_FILE
+    sed -i 's/\$Ros_distro\$/'${some_ros_distro^}'/g' $SED_FILE
+    sed -i 's/\$NAME\$/'${repo_name}'/g' $SED_FILE
+    sed -i 's/\$NAMESPACE\$/'${repo_namespace}'/g' $SED_FILE
+  done
+
+  # append to readme files
   tail -n +$CUT_LINE README.md >> $TMP_FILE
   mv $TMP_FILE README.md
 
+  cat ${PACKAGE_TEMPLATES}/_append_to_README_build_status.md >> $TMP_FILE_2
+
+  cat $TMP_FILE_2 >> .github/workflows/README.md
+  rm $TMP_FILE_2
+
   echo ""
-  echo -e "${TERMINAL_COLOR_BLUE}Added setup for ros distro '${some_ros_distro} on branch ${some_branch}.${TERMINAL_COLOR_NC}"
+  echo -e "${TERMINAL_COLOR_BLUE}Added setup for ros distro '${RAW_TERMINAL_COLOR_BROWN}${some_ros_distro}${TERMINAL_COLOR_BLUE}' on branch '${RAW_TERMINAL_COLOR_BROWN}${some_branch}${TERMINAL_COLOR_BLUE}'.${TERMINAL_COLOR_NC}"
   echo ""
 
   read -p "${RAW_TERMINAL_COLOR_BROWN}Do you want to configure ci for another ros distro?${RAW_TERMINAL_COLOR_NC} (yes/no) [no] " another
@@ -170,17 +189,22 @@ while true; do
   fi
 
   echo "" > $TMP_FILE
+  echo "" > $TMP_FILE_2
 
   read -p "${RAW_TERMINAL_COLOR_BROWN}Name of the ROS distro? ${RAW_TERMINAL_COLOR_NC}" some_ros_distro
   read -p "${RAW_TERMINAL_COLOR_BROWN}Name of the repository branch?: ${RAW_TERMINAL_COLOR_NC}" some_branch
 done
 
 # Setting up formatting
-cp -n ${PACKAGE_TEMPLATES}/.clang-format .
-cp -n ${PACKAGE_TEMPLATES}/.pre-commit-config.yaml .
-pre-commit install
-pre-commit autoupdate
+read -p "${RAW_TERMINAL_COLOR_BROWN}Do you want to setup formatting using pre-commit?${RAW_TERMINAL_COLOR_NC} (yes/no) [no]: " formatting
+formatting=${formatting:="no"}
 
+if  [[ "$formatting" == "yes" ]]; then
+  cp -n ${PACKAGE_TEMPLATES}/.clang-format .
+  cp -n ${PACKAGE_TEMPLATES}/.pre-commit-config.yaml .
+  pre-commit install
+  pre-commit autoupdate
+fi
 
 echo ""
 echo -e "${TERMINAL_COLOR_BLUE}FINISHED: Please check the generated file and setup package names.${TERMINAL_COLOR_NC}"
